refactor: Rename `kraft pkg prune` to `kraft pkg rm`
This is more inline with existing subcommands for `kraft`.  Also rename
the package manager's `Prune` method to `Delete` since "prune" implies
all and the method accepts query parameters to allow for selecting all
or a subset.
