+++ cmd/kraft/pkg/pkg.go
@@ -25,9 +25,9 @@ import (
 	"kraftkit.sh/unikraft/app"
 
 	"kraftkit.sh/cmd/kraft/pkg/list"
-	"kraftkit.sh/cmd/kraft/pkg/prune"
 	"kraftkit.sh/cmd/kraft/pkg/pull"
 	"kraftkit.sh/cmd/kraft/pkg/push"
+	"kraftkit.sh/cmd/kraft/pkg/rm"
 	"kraftkit.sh/cmd/kraft/pkg/source"
 	"kraftkit.sh/cmd/kraft/pkg/unsource"
 	"kraftkit.sh/cmd/kraft/pkg/update"
@@ -81,10 +81,10 @@ func New() *cobra.Command {
 	cmd.AddCommand(list.New())
 	cmd.AddCommand(pull.New())
 	cmd.AddCommand(push.New())
+	cmd.AddCommand(rm.New())
 	cmd.AddCommand(source.New())
 	cmd.AddCommand(unsource.New())
 	cmd.AddCommand(update.New())
-	cmd.AddCommand(prune.New())
 
 	return cmd
 }
