+++ packmanager/umbrella.go
@@ -95,13 +95,13 @@ func (u UmbrellaManager) AddSource(ctx context.Context, source string) error {
 	return nil
 }
 
-func (u UmbrellaManager) Prune(ctx context.Context, qopts ...QueryOption) error {
+func (u UmbrellaManager) Delete(ctx context.Context, qopts ...QueryOption) error {
 	var errs []error
 	for _, manager := range u.packageManagers {
 		log.G(ctx).WithFields(logrus.Fields{
 			"format": manager.Format(),
 		}).Tracef("pruning")
-		err := manager.Prune(ctx, qopts...)
+		err := manager.Delete(ctx, qopts...)
 		if err != nil {
 			if strings.Contains(err.Error(), "package not found") {
 				errs = append(errs, err)
